backend supports real time notifcations. need to add/fix tests
thank you galco
