+++ backend/app.py
@@ -1,13 +1,12 @@
 from flask import Flask
-from flask_socketio import SocketIO
+
 # from service_layer.routes import api_routes  # Import API routes from service_layer
-from __init__ import create_app
+from backend import create_app
 import logging
 
 logger = logging.getLogger('myapp')
 
 app = create_app()
-socketio = SocketIO(app)
 
 # Register API routes from service_layer
 #app.register_blueprint(api_routes)
